requirements.txt
# Core dependencies
overrides>=7.4.0
class_registry
griffe
duckduckgo_search
loguru
langchain_text_splitters
trafilatura
tqdm
python-dotenv
dill
botocore
boto3
urllib3
google-generativeai>=0.3.0
litellm[caching]
mem0ai[graph]
setuptools
langchain-memgraph
#qdrant-client
#neo4j
#langchain




# Development dependencies
pytest
gradio